backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pyjwt==2.6.0
cryptography==40.0.2
requests==2.30.0
sqlalchemy==1.4.40
passlib==1.7.4
python-dotenv==1.0.0
gunicorn==20.1.0
redis==4.5.1
celery==5.2.7
click==8.1.3
pycryptodomex==3.17
